fix(infra): scope jsdom setup; tighten CSP img-src; trim WHAT comments
Vitest's inline test.projects with extends:true had no effect in
vitest 2.1.9 — workspace isolation only activates via a workspace
config file. Moved project definitions to vitest.workspace.ts so node
and jsdom environments are truly isolated: DOM matchers are now loaded
only in the jsdom project (setupFiles on the jsdom project only), and
the node project never inherits @testing-library/jest-dom.

CSP img-src tightened from 'https:' to GitHub-owned hosts
(*.githubusercontent.com, github.com) to mirror the sanitizer's
image-src allowlist after the readme.ts fixes.

Trimmed WHAT comments per project convention: ReadmeCache html/etag
field comments (field names already say it), and the .prose-pluma
header in layout.css.
