+++ vitest.workspace.ts
@@ -1,0 +1,55 @@
+import { defineWorkspace } from 'vitest/config';
+// vite-plugin-svelte-4 is @sveltejs/vite-plugin-svelte@^4 aliased for Vite 5 (vitest's vite-node)
+import { svelte } from 'vite-plugin-svelte-4';
+import { svelteTesting } from '@testing-library/svelte/vite';
+import { fileURLToPath } from 'node:url';
+
+const stub = (p: string) => fileURLToPath(new URL(`./src/__test__/stubs/${p}`, import.meta.url));
+
+const sharedAliases = {
+	'$env/dynamic/private': stub('env-private.ts'),
+	'$env/dynamic/public': stub('env-public.ts'),
+	'$app/environment': stub('app-environment.ts'),
+	$lib: fileURLToPath(new URL('./src/lib', import.meta.url))
+};
+
+const sharedEnv = {
+	JWT_SECRET: 'test-secret-please-do-not-use-in-production-1234567890',
+	ROOT_EMAIL: 'root@pluma.local',
+	ORIGIN: 'http://localhost:5173',
+	NODE_ENV: 'test'
+};
+
+export default defineWorkspace([
+	{
+		// Node project: server/unit tests only — no DOM matchers loaded.
+		plugins: [svelte()],
+		resolve: { alias: sharedAliases },
+		test: {
+			name: 'node',
+			env: sharedEnv,
+			include: ['src/**/*.{test,spec}.{js,ts}'],
+			exclude: [
+				'src/**/*.svelte.{test,spec}.{js,ts}',
+				'src/**/*-render.{test,spec}.{js,ts}'
+			],
+			environment: 'node'
+		}
+	},
+	{
+		// jsdom project: component render tests — @testing-library/jest-dom matchers
+		// are loaded via setupFiles; svelteTesting() ensures browser resolve conditions.
+		plugins: [svelte(), svelteTesting()],
+		resolve: { alias: sharedAliases },
+		test: {
+			name: 'jsdom',
+			env: sharedEnv,
+			include: [
+				'src/**/*.svelte.{test,spec}.{js,ts}',
+				'src/**/*-render.{test,spec}.{js,ts}'
+			],
+			environment: 'jsdom',
+			setupFiles: ['src/__test__/jsdom-setup.ts']
+		}
+	}
+]);
